+++ CoreNotesView/src/au/gov/asd/tac/constellation/views/notes/utilities/DateTimePicker.java
@@ -18,11 +18,24 @@ package au.gov.asd.tac.constellation.views.notes.utilities;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.time.chrono.Chronology;
+import java.time.format.DateTimeFormatter;
+import java.util.Locale;
+import javafx.geometry.Insets;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
 import javafx.scene.control.Spinner;
+import javafx.scene.layout.Border;
+import javafx.scene.layout.BorderStroke;
+import javafx.scene.layout.BorderStrokeStyle;
+import javafx.scene.layout.BorderWidths;
+import javafx.scene.layout.CornerRadii;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
+import javafx.scene.paint.Color;
+import javafx.util.StringConverter;
+import org.apache.commons.lang3.StringUtils;
+
 
 /**
  * A DateTime Selector that is in JavaFX instead of Swing.
@@ -48,6 +61,24 @@ public class DateTimePicker {
 
     public DateTimePicker(final boolean from) {
         dateTimePane = new Pane();
+        datePicker.setChronology(Chronology.ofLocale(Locale.ENGLISH));
+
+        datePicker.setConverter(new StringConverter<LocalDate>() {
+            final String pattern = "yyyy-MM-dd";
+            final DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern(pattern);
+
+            @Override
+            public String toString(final LocalDate object) {
+                return object != null ? dateFormatter.format(object) : "";
+            }
+
+            @Override
+            public LocalDate fromString(final String string) {
+                return StringUtils.isNotBlank(string) ? LocalDate.parse(string, dateFormatter) : null;
+            }
+
+        });
+
         this.from = from;
         dateTimePane.getChildren().add(mainGridPane);
 
@@ -64,7 +95,7 @@ public class DateTimePicker {
 
         datePickerGridPane.add(datePickerLabel, 0, 0);
         datePickerGridPane.add(datePicker, 1, 0);
-
+        datePicker.setMaxWidth(150);
         final GridPane timePickerGrid = new GridPane();
         final Label hourLabel = new Label("Hour");
         final Label minLabel = new Label("Minute");
@@ -92,12 +123,23 @@ public class DateTimePicker {
 
         mainGridPane.add(datePickerGridPane, 0, 0);
         mainGridPane.add(timePickerGrid, 0, 1);
+        mainGridPane.setPadding(new Insets(1, 1, 1, 1));
+
+        dateTimePane.setBorder(new Border(new BorderStroke(Color.WHITE, BorderStrokeStyle.SOLID, new CornerRadii(3.0), BorderWidths.DEFAULT)));
+    }
+
+    public void disableControls(final boolean disable) {
+        datePicker.setDisable(disable);
+        hourPicker.setDisable(disable);
+        minPicker.setDisable(disable);
+        secPicker.setDisable(disable);
     }
 
 
     public Pane getPane() {
         return dateTimePane;
     }
+
 
     /**
      * Sets the current date time to whatever the current local time is